refactor: id type 변경

+++ backend/src/main/java/com/ssafy/arttab/artwork/ArtworkRepository.java
@@ -2,6 +2,6 @@ package com.ssafy.arttab.artwork;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArtworkRepository extends JpaRepository<Artwork, Integer> {
+public interface ArtworkRepository extends JpaRepository<Artwork, Long> {
 
 }
